Got OAuth flow working with Google photos API.
The basic OAuth flow is working by asking user to visit Google to grant
the application permission to their Google Photo library. After the
user grants access, then the refresh token is fetched an stored in a
password manager called keyring-rs. Afterwards, the OAuth flow is not
required and the application will be able to access the Google Photo api
with that token until it expires or the user revokes access.

The application fetches the latest photo and stores it in the current
directory that it was run in.

There are some basic tests written to check the basic OAuth flow.

Next will work on implementing the photo download so that it can handle
more than the latest photo. I've also begun thinking about how to index
photos/media for synchronization, but this seemed like a good time to
stop as the basic flow is working end to end.

More error handling will be required, but for now I'm unwrapping like it
is Christmas.