Refactor the media downloader into a separate Struct
Move the code for fetching media metadata and binary into the library
module under the name MediaFetcher. This has two methods for fetching
the metadata and then writing it out to disk. One implication after
making this change was to make the error type more generic. So that
was also refactored into an Error enum that wraps the HTTP request and
serialization failures. The last change made was to use Structs instead
of the raw Json values, which takes advantage of the Type system. Overall
think the code is nicer to read and simplified a bit.