[Test] Add integration test to check MediaFetcher can write albums
Made the directory where files are written parameterizable and use a
temp file to check MediaFetcher's ability to write out the
content of an album. The program currently writes to the CWD but will
make this an option to the CLI. The test makes sure the correct endpoint
is called and the contents that are returned are written to disk.

Unfortunately the tempfile doesn't give a directory, so the temp file
that is created by the test isn't cleaned up. Ideally should change this
around to some kind of temp directory instead.