[MediaFetcher] More WIP to handle pagination
Finished wiring up the pageToken to another request and merging the
media item results back into the return value. Looks like it's working,
so next will try and fetch media items for a longer time period until all
have been retrieved to see if it would work to handle downloading
an entire library or some subset of one.

This commit breaks some of the tests until can resolve an issue with
httpmock or re-write the test.