[MediaFetcher] Fix broken pagination test
The pagination test had been broken because of a misunderstanding I had
regarding how the httpmock matched requests. Apparently since the first
mock was a subset of the second, it was matched first in both cases
despite the request missing the query parameter. The fix provided by
the library author was to add a match that made sure the query param
didn't exist on the request.