[MediaFetcher] Fetch the requested number of media
The page size is the maximum number of items that may be returned by Google,
so the MediaFetcher could fetch metadata for more than the requested number
of media if there are subsequent requests. When writing out media it will
only fetch up to the specified number.

Also playing around with the output so that it's a little more helpful. Now
prints out the year, month, and day of the media item being downloaded.