[WIP] Work on functionality to resume progress
Want to add the functionality to resume progress on fetching a library.
This will find the latest date that was downloaded and then resume downloading
from that day. If a file already exists for a media item, it will skip
over it instead of re-writing.

Leaving code in non-compilable state for now.