[lib] Add method to find recent year, month, day
Add a new method to the library that takes a base path expecting a dir
structure like photos/yyyy/mm/dd and will return the most recent year,
month, and day based on the most recent directories found therein. This
will be used to begin fetching only the most recent number of photos
after this point in time in hopes that this will be good enough to
resume downloading if interrupted.